+++ internal/service/tag.go
@@ -7,29 +7,29 @@ import (
 
 type CountTagRequest struct {
 	Name  string `form:"name" binding:"max=100"`
-	State int8   `form:"state" binding:"omitempty,oneof=-1 0 1"`
+	State int    `form:"state" binding:"omitempty,oneof=-1 0 1"`
 }
 
 type TagListRequest struct {
 	Name  string `form:"name" binding:"max=100"`
-	State int8   `form:"state,default=-1" binding:"omitempty,oneof=-1 0 1"`
+	State int    `form:"state,default=-1" binding:"omitempty,oneof=-1 0 1"`
 }
 
 type CreateTagRequest struct {
 	Name      string `form:"name" binding:"required,min=2,max=100"`
 	CreatedBy string `form:"created_by" binding:"required,min=2,max=100"`
-	State     int8   `form:"state,default=1" binding:"omitempty,oneof=0 1"`
+	State     int    `form:"state,default=1" binding:"omitempty,oneof=0 1"`
 }
 
 type UpdateTagRequest struct {
-	ID         uint32 `form:"id" binding:"required,gte=1"`
+	ID         int    `form:"id" binding:"required,gte=1"`
 	Name       string `form:"name" binding:"max=100"`
-	State      int8   `form:"state,default=-1" binding:"omitempty,oneof=-1 0 1"`
+	State      int    `form:"state,default=-1" binding:"omitempty,oneof=-1 0 1"`
 	ModifiedBy string `json:"modified_by" form:"modified_by" binding:"required,min=2,max=100"`
 }
 
 type DeleteTagRequest struct {
-	ID uint32 `form:"id" binding:"required,gte=1"`
+	ID int `form:"id" binding:"required,gte=1"`
 }
 
 func (svc *Service) CountTag(param *CountTagRequest) (int, error) {
